dags/daily_employees/requirements.txt
# Dependencies for daily_csv_to_iceberg DAG
# Version: 1.0.0
# Last updated: 2025-11-10

# Spark and Iceberg
pyspark==3.5.3
pyiceberg==0.10.0
pyarrow>=17.0.0

# Additional dependencies for date handling
python-dateutil>=2.9.0
